requirements.txt
Flask
Pillow
gunicorn